Implement Linux/MacOS/Windows process managers
