+++ src/Microsoft.DotNet.XHarness.Common/Execution/IMacOSProcessManager.cs
@@ -7,10 +7,9 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Microsoft.DotNet.XHarness.Common.Execution;
 using Microsoft.DotNet.XHarness.Common.Logging;
 
-namespace Microsoft.DotNet.XHarness.iOS.Shared.Execution
+namespace Microsoft.DotNet.XHarness.Common.Execution
 {
     public interface IMacOSProcessManager : IProcessManager
     {
