+++ src/Microsoft.DotNet.XHarness.Common/Execution/IProcessManager.cs
@@ -1,7 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
-#nullable enable
 
 using System;
 using System.Collections.Generic;
@@ -10,9 +9,9 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Common.Logging;
 
+#nullable enable
 namespace Microsoft.DotNet.XHarness.Common.Execution
 {
-
     public class ProcessExecutionResult
     {
         public bool TimedOut { get; set; }
@@ -20,13 +19,47 @@ namespace Microsoft.DotNet.XHarness.Common.Execution
         public bool Succeeded => !TimedOut && ExitCode == 0;
     }
 
-    // interface that helps to manage the different processes in the app.
+    /// <summary>
+    /// Interface that helps to manage processes
+    /// </summary>
     public interface IProcessManager
     {
-        Task<ProcessExecutionResult> ExecuteCommandAsync(string filename, IList<string> args, ILog log, TimeSpan timeout, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null);
-        Task<ProcessExecutionResult> ExecuteCommandAsync(string filename, IList<string> args, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan timeout, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null);
-        Task<ProcessExecutionResult> RunAsync(Process process, ILog log, TimeSpan? timeout = null, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
-        Task<ProcessExecutionResult> RunAsync(Process process, ILog log, ILog stdoutLog, ILog stderrLog, TimeSpan? timeout = null, Dictionary<string, string>? environmentVariables = null, CancellationToken? cancellationToken = null, bool? diagnostics = null);
+        Task<ProcessExecutionResult> ExecuteCommandAsync(
+            string filename,
+            IList<string> args,
+            ILog log,
+            TimeSpan timeout,
+            Dictionary<string, string>? environmentVariables = null,
+            CancellationToken? cancellationToken = null);
+
+        Task<ProcessExecutionResult> ExecuteCommandAsync(
+            string filename,
+            IList<string> args,
+            ILog log,
+            ILog stdoutLog,
+            ILog stderrLog,
+            TimeSpan timeout,
+            Dictionary<string, string>? environmentVariables = null,
+            CancellationToken? cancellationToken = null);
+
+        Task<ProcessExecutionResult> RunAsync(
+            Process process,
+            ILog log,
+            TimeSpan? timeout = null,
+            Dictionary<string, string>? environmentVariables = null,
+            CancellationToken? cancellationToken = null,
+            bool? diagnostics = null);
+
+        Task<ProcessExecutionResult> RunAsync(
+            Process process,
+            ILog log,
+            ILog stdoutLog,
+            ILog stderrLog,
+            TimeSpan? timeout = null,
+            Dictionary<string, string>? environmentVariables = null,
+            CancellationToken? cancellationToken = null,
+            bool? diagnostics = null);
+
         Task KillTreeAsync(Process process, ILog log, bool? diagnostics = true);
         Task KillTreeAsync(int pid, ILog log, bool? diagnostics = true);
     }
