+++ src/Microsoft.DotNet.XHarness.Common/Execution/WindowsProcessManager.cs
@@ -1,0 +1,26 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Diagnostics;
+using Microsoft.DotNet.XHarness.Common.Logging;
+
+namespace Microsoft.DotNet.XHarness.Common.Execution
+{
+    public class WindowsProcessManager : ProcessManager
+    {
+        // TODO: Maybe we don't need this because we can kill the whole tree for netcoreapp3.1
+        protected override List<int> GetChildrenPS(ILog log, int pid) => new List<int>();
+
+        protected override int Kill(int pid, int sig)
+        {
+#if NETCOREAPP3_1
+            Process.GetProcessById(pid).Kill(entireProcessTree: true);
+#else
+            Process.GetProcessById(pid).Kill();
+#endif
+            return 0;
+        }
+    }
+}
