+++ src/Microsoft.DotNet.XHarness.Common/Logging/ConsoleLog.cs
@@ -6,7 +6,7 @@ using System;
 using System.IO;
 using System.Text;
 
-namespace Microsoft.DotNet.XHarness.iOS.Shared.Logging
+namespace Microsoft.DotNet.XHarness.Common.Logging
 {
     /// <summary>
     /// A log that writes to standard output
