+++ src/Microsoft.DotNet.XHarness.Common/Logging/ILog.cs
@@ -3,12 +3,10 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.IO;
 using System.Text;
 
 namespace Microsoft.DotNet.XHarness.Common.Logging
 {
-
     public interface ILog : IDisposable
     {
         string? Description { get; set; }
@@ -17,18 +15,7 @@ namespace Microsoft.DotNet.XHarness.Common.Logging
         void Write(byte[] buffer, int offset, int count);
         void Write(string value);
         void WriteLine(string value);
-        void WriteLine(StringBuilder value);
         void WriteLine(string format, params object[] args);
         void Flush();
-    }
-
-    public interface IReadableLog : ILog
-    {
-        StreamReader GetReader();
-    }
-
-    public interface IFileBackedLog : IReadableLog
-    {
-        string FullPath { get; }
     }
 }
